+++ lib/constants/constants.dart
@@ -4,5 +4,6 @@ const bool DEBUG = true;
 
 const STATUS_TODO = 'todo';
 const STATUS_REVIEW = 'review';
-const STATUS_VERIFIED = 'verified';
-const STATUS_REVISION = 'revision';
+const STATUS_ACCEPTED = 'accepted';
+const STATUS_REJECTED = 'rejected';
+const STATUS_EXPIRED = 'expired';