+++ CHANGELOG.md
@@ -62,3 +62,6 @@
 - Handling login error
 - Sebelum upload asset, maker harus upload e-sign
 - Add loading indicator when load asset from server (selain todo)
+# 28/12/2023
+- proteksi report ketika status bukan todo
+- reset password